requirements.txt
lxml==4.6.3
PyExecJS==1.5.1